API-2218 moved PlayBindings into apigateway package
